+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+  - Added BeGreaterThan and BeLessThan assertions to Should.
+
 ## 3.1.1 (October 29, 2014)
   - Fix Skipped and Pending
   - Fix output format on non-US systems