+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
   - Added BeGreaterThan and BeLessThan assertions to Should.
   - Add -Quiet parameter for Invoke-Pester that disables the output written to screen by Write-Host [GH-223]
+  - Fix Error output for TestDrive [GH-232]
 
 ## 3.1.1 (October 29, 2014)
   - Fix Skipped and Pending